Sort parameters array before use in request
Sometimes when PDO parameters are provided to the `exec()` method, the array may have positional arguments specified out of order with the statement. For instance:

```
$pdo = new \Crate\PDO\PDO('crate:127.0.0.1:4200');
$stmt = $pdo->prepare('UPDATE schema.table SET colA = :colA WHERE colB = :colB');
$stmt->execute(['colB' => 'valB', 'colA' => 'valA']);
```

Without the above line, the above code would lead to the core server choking since the parameter bag will be an object rather than an array:

```
SQLActionException[
    Failed to parse source [{
        "stmt" : "UPDATE schema.table SET colA = ? WHERE colB = ?",
        "args" : {"1" : "valB", "0" : "valA"}
    }]
]
```

Providing the arguments out of order in this way is not necessarily malicious as there are no prescriptions or restrictions that prevent it from happening.