Replace Coveralls by Codecov
We found that Coveralls is not able to merge coverage reports from
different job runs. Since supporting PHP8, we have different code
branches in order to support the old and the new PDO interface.

So, it is crucial to merge coverage reports from two job runs, one
for PHP7.4 and another for PHP8.